문&오브젝트 수정
코드 분리
